+++ BlueprintUICommonControls/Sources/AttributedLabel+Environment.swift
@@ -29,7 +29,16 @@ class DefaultURLHandler: NullURLHandler {
 }
 
 public struct URLHandlerEnvironmentKey: EnvironmentKey {
-    public static var defaultValue: URLHandler = DefaultURLHandler()
+    public static let defaultValue: URLHandler = {
+        // This is our best guess for "is this executable an extension?"
+        if let _ = Bundle.main.infoDictionary["NSExtension"] {
+            return NullURLHandler()
+        } else if Bundle.main.bundlePath.hasSuffix(".appex") {
+            return NullURLHandler()
+        } else {
+            return DefaultURLHandler()
+        }
+    }()
 }
 
 extension Environment {