+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Updated minimum deployment target from iOS 12 to iOS 14.
 - `URLHandlerEnvironmentKey.defaultValue` is now settable.
+- Marks pod as `APPLICATION_EXTENSION_API_ONLY`
 
 ### Deprecated